Replace nose with pytest, refactor to more methods
This adds some tests that are run by pytest.
There's a small refactor to break up the main logic into a few methods.
I don't love it, and it'll need another pass, but it's helping expose
the dependencies between the steps.

I also added a check for "new" items that are sold out. This happens
when things sell out but aren't removed from the main grid.